src/infrastructure/config/requirements.txt
google-adk>=0.4.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
aiofiles>=23.2.1
openai>=1.3.0
pyyaml>=6.0 